Use more precise Azul Zulu release names
Azul is distributing their Zulu OpenJDK builds with their own version, which doesn't directly correspond to a single OpenJDK build because there could be multiple Zulu releases per OpenJDK release.

The release names didn't reflect this, so it was unclear which version of the Zulu OpenJDK was being installed.

Existing users should not be affected by this change. They will only see the "new" versions of Azul Zulu in the output of `asdf list-all java`.