System Java now works on macOS
Fixes https://github.com/halcyon/asdf-java/issues/132

Added special case for macOS in all 4 set-java-home scripts. Also created tests to validate this. Unfortunately xonsh does not seem to have a way to exit with other status then 0, so this will not report errors (yet).

Updated the README.md to contain information on how to use the set-java-home scripts.

Cloned mstksg/get-package and released delgurth/get-package@v3 because it was using linuxbrew on Ubuntu and because I needed the apt-update (not released by mstksg) and -y flag in the apt-get install (when running github action on https://github.com/nektos/act)

added .editorconfig because of mixed indent styles
added .gitignore to ignore Intellij directory

asdf/asdf.sh needs $PWD in front as soon as you change directories

added tests for JAVA_HOME setting
added tests for macOS JAVA_HOME setting in case version is set to system

Unfortunately the tests are not exactly as it would be on a system because of the way github actions spawn shells. Have to call the bash function _asdf_java_prompt_command and the fish & zsh function asdf_update_java_home to actually update the JAVA_HOME.