Liberica now also has macOS Integration support
With the help of BellLabs (https://github.com/bell-sw/Liberica/issues/42) we now also have Liberica macOS integration support. It doesn't work for all versions we can install, so it logs a warning in case it's not possible to use their "trick".

For this to work we depend on @joschi java-metadata project. This project provides the shasum (and with that the check if the .pkg version exists).

Other things done:
- Fail if commands fail
- Fixed unzip check
- All error messages go to stderr
- Aligned indent to 4 spaces
- Removed all echo's in favor of printf in line with https://github.com/asdf-vm/asdf/pull/806
- Use the bash [[ conditions